% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_orientation.R
\name{assign_orientation}
\alias{assign_orientation}
\title{Assign orientation}
\usage{
assign_orientation(
  x,
  formula = vessel + cruise + haul ~ cdepth,
  resid.col,
  resid.cut,
  output.name = "orientation"
)
}
\arguments{
\item{x}{Data frame variables used to assign orientation.}

\item{formula}{Function used to cast data frame, where the depth column used for casting is the response.}

\item{resid.col}{Name of the column containing residual values.}

\item{resid.cut}{Threshold value for assigning residuals.}

\item{output.name}{Name of the orientation column. Default = "orientation"}
}
\value{
Returns `x` with an new column indicating the assigned orientation.
}
\description{
Function to assign orientation based on visual inspection of residual plots. Assigns orientation based on residuals for multiple depth bins,
simultaneously.
}
\author{
S.K. Rohan \email{skrohan@uw.edu}
}
