% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_od.R
\name{calculate_od}
\alias{calculate_od}
\title{Calculate optical depth}
\usage{
calculate_od(x)
}
\arguments{
\item{x}{A vector of light measurements to be converted to optical depth.}
}
\value{
Optical depth
}
\description{
\code{calculate_od} converts light measurements (photon flux density, etc.) to optical depth (Kirk 2011) for each observation, relative to the highest observed light level.
}
\references{
Kirk, J.T.O. 2011. Light and photosynthesis in aquatic ecosystems. In 3rd edition. Cambridge University Press, New York.
}
\author{
S.K. Rohan \email{skrohan@uw.edu}
}
