% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cast_wrapper}
\alias{cast_wrapper}
\title{Applies a function to all casts in a data frame}
\usage{
cast_wrapper(x, id.col, FUN, min.rows = 4, silent = TRUE, ...)
}
\arguments{
\item{x}{Data frame}

\item{id.col}{Name of column containing unique identifier for variables (i.e. haul, vessel, cruise, up/downcast) apply function.}

\item{FUN}{Function to apply}

\item{min.rows}{Minimum number of rows necessary to apply functions.}

\item{silent}{Should messages be printed to console while the function is running?}

\item{...}{Additional arguments passed to internal functions.}
}
\description{
For use processing AOPs from AFSC/RACE/GAP data structure.
}
