% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_light.R
\name{convert_light}
\alias{convert_light}
\title{Convert archival tag light to PFD}
\usage{
convert_light(x, convert.method = "kotwicki")
}
\arguments{
\item{x}{Archival tag light measurement in relative units (~25-225)}

\item{convert.method}{Character vector indicating which method to use. Options: kotwicki, wc. Default: kotwicki}
}
\description{
\code{convert_light} converts light measurements from relative units recorded by a Wildlife Computers TDR-Mk9 archival tag equipped with a Hamatsu S2387 silicone photodiode to quantum units based on the equation of Kotwicki et al. (2009), or blue light intensity reported by Wildlife Computers.
}
\references{
Kotwicki, S., Robertis, A., von Szalay, P., and Towler, R. 2009. The effect of light intensity on the availability of walleye pollock (Theragra chalcogramma) to bottom trawl and acoustic surveys. Can. J. Fish. Aquat. Sci. 66(6): 983–994. doi:10.1139/f09-055.
}
\author{
Sean Rohan \email{sean.rohan@noaa.gov}
}
