% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{csv_rbind}
\alias{csv_rbind}
\title{Open and bind rows of csv to data frame}
\usage{
csv_rbind(directory, string)
}
\arguments{
\item{directory}{Path to directory where csv files are stored.}

\item{string}{Unique pattern used in the csv files.}
}
\description{
For use with AFSC/RACE/GAP data structure.
}
