% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{energy_to_quanta}
\alias{energy_to_quanta}
\title{Convert radiometric energy to photon flux density}
\usage{
energy_to_quanta(x, wavelength)
}
\arguments{
\item{x}{Numeric vector. Energy for a wavelength in units of watts.}

\item{wavelength}{Numeric vector.Wavelength in nanometers.}
}
\value{
Numeric vector of energy in photon flux density.
}
\description{
Convert radiometric energy in watts to micromoles of photons per meter squared per second.
}
