% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_stepwise.R
\name{filter_stepwise}
\alias{filter_stepwise}
\title{Bin and filter light measurements}
\usage{
filter_stepwise(cast.data, light.col, depth.col, bin.size = 2, bin.gap = 6,
  agg.fun, ...)
}
\arguments{
\item{cast.data}{Data frame containing light measurements and depth.}

\item{bin.size}{The size of the depth bin used for aggregation. Default = 2.}

\item{bin.gap}{The maximum size of data gap before a profile is considered to not meet continuity standards. Units are in units of depth, not the number of bins. Default = 6.}

\item{agg.fun}{Function used to aggregate light measurements for each depth bin.}

\item{...}{Additional arguments passed to findInterval function for binning light measurements}
}
\value{
Data frame with light by depth bin and continuity grade (1 = Good, -999 = Bad)
}
\description{
\code{filter_stepwise} aggregates light measurements from a single cast into depth bins using a specificed function (e.g. 'median'), removes light measurements using a stepwise algorithm, and assigns data continuity grades based on a threshold criteria.
}
