% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_mk9_offset}
\alias{find_mk9_offset}
\title{Find and apply time offset to Mk9 light data}
\usage{
find_mk9_offset(light, mbt, try.offsets = seq(-8, 8, 0.5), results.file = NULL)
}
\arguments{
\item{light}{Data frame containing Mk9 data. Must include columns: ldate_time (POSIXct), ldepth (numeric)}

\item{mbt}{Data frame containing MBT data. Must include columns: date_time (POSIXct), depth (numeric)}

\item{try.offsets}{A vector of offsets to try. Default is seq(-8,8,0.5)}

\item{results.file}{Character vector specifying the filepath where information about the offset and correlation between Mk9 and MBT depths are stored.}
}
\value{
The input light data frame with date_time adjusted according to the offset.
}
\description{
For use processing AOPs from AFSC/RACE/GAP data structure.
}
\author{
Sean Rohan \email{sean.rohan@noaa.gov}
}
