% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optical_depth.R
\name{find_optical_depth}
\alias{find_optical_depth}
\title{Depth of optical depth}
\usage{
find_optical_depth(
  x,
  od.col = "optical_depth",
  depth.col = "cdepth",
  target.od = 2.302585,
  with.input = F,
  return.col = F
)
}
\arguments{
\item{x}{Data frame containing optical depth and bin depth for a single cast.}

\item{od.col}{Name of the data frame column containing optical depth}

\item{depth.col}{Name of the data frame column containing depths}

\item{target.od}{Optical depth value of interest.}

\item{with.input}{Logical. Should the input data frame be returned with the optical depth? By default, false.}

\item{return.col}{Optional. Name of the column for output. Default = F.}
}
\value{
Returns either a numerical vector with the optical depth for the cast, or a data frame with the input data and the optical depth data for the cast. Optical depths exceeding the maximum depth of the cast are returned as an NA.
}
\description{
Function to estimate the depth of a user-specified optical depth.
}
\author{
S.K. Rohan \email{skrohan@uw.edu}
}
