% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_mean.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Geometric mean}
\usage{
geometric.mean(x, na.rm = T)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical. Should NA values be omitted from calculation? Default = T}
}
\description{
Calculates the geometric mean of a numeric vector.
}
