% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_as2.R
\name{loess.as2}
\alias{loess.as2}
\title{Local polynomial regression with automatic span selection and user-defined span bounds}
\usage{
loess.as2(
  x,
  y,
  degree = 1,
  criterion = c("aicc", "gcv"),
  family = c("gaussian", "symmetric"),
  min.bins = 3,
  user.span = NULL,
  plot = FALSE,
  ...
)
}
\arguments{
\item{x}{as described for \code{fANCOVA::loess.as}}

\item{y}{as described for \code{fANCOVA::loess.as}}

\item{degree}{as described for \code{fANCOVA::loess.as}}

\item{criterion}{as described for \code{fANCOVA::loess.as}}

\item{family}{as described for \code{fANCOVA::loess.as}}

\item{min.bins}{Minimum number of depth bins over which a span can smooth. Default = 3}

\item{user.span}{as described for \code{fANCOVA::loess.as}}

\item{plot}{as described for \code{fANCOVA::loess.as}}

\item{...}{as described for \code{fANCOVA::loess.as}}
}
\description{
\code{loess.as2} fits a local polynomial gregression with automatic smoothing paramater selection, but has been modified to facilitate user-specified dynamic span bounds based on the number of observations using the \code{min.bins} argument. The function description and code have been modified from \code{fANCOVA::loess.as} v0.5-1 (Wang 2010).
}
\references{
Cleveland, W. S. (1979) Robust locally weighted regression and smoothing scatterplots. Journal of the American Statistical Association. 74, 829–836.

Golub, G., Heath, M. and Wahba, G. (1979). Generalized cross validation as a method for choosing a good ridge parameter. Technometrics. 21, 215–224.

Hurvich, C.M., Simonoff, J.S., and Tsai, C.L. (1998), Smoothing Parameter Selection in Nonparametric Regression Using an Improved Akaike Information Criterion. Journal of the Royal Statistical Society B. 60, 271–293.

Xiao-Feng Wang (2010). fANCOVA: Nonparametric Analysis of Covariance. R package version 0.5-1. https://CRAN.R-project.org/package=fANCOVA
}
\author{
Sean Rohan \email{sean.rohan@noaa.gov}
}
