% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aop_ncdf.R
\name{make_aop_ncdf}
\alias{make_aop_ncdf}
\title{Make netCDF files}
\usage{
make_aop_ncdf(
  region,
  year,
  global_attributes = list(title =
    "Apparent Optical Properties from AFSC 2021 EBS Shelf and NBS Bottom Trawl Surveys",
    references =
    "Rohan, S.K., Kotwicki, S., Kearney, K.A., Schulien, J.A., Laman, E.A., Cokelet, E.D., Beauchamp, D.A., Britt, L.L., Aydin, K.Y., & Zador, S.G. (2021). Using bottom trawls to monitor subsurface water clarity in marine ecosystems. Progress in Oceanography, 194, 102554. https://doi.org/10.1016/j.pocean.2021.102554",
    id = "https://doi.org/10.5281/zenodo.3688864", cdm_data_type = "Point", cruise =
    "2021 Eastern Bering Sea Continental Shelf and Northern Bering Sea Bottom-Trawl Surveys",
    
     institution = "NOAA Alaska Fisheries Science Center", contributor_name =
    "Rebecca Haehn, Ned Laman", creator_name = "Sean Rohan", creator_institution =
    "NOAA Alaska Fisheries Science Center", creator_email = "sean.rohan@noaa.gov",
    publisher = "NOAA Alaska Fisheries Science Center", publisher_type = "institution",
    publisher_url =
    "https://www.fisheries.noaa.gov/about/alaska-fisheries-science-center",
    geospatial_bounds_crs = "EPSG:4326", license =
    "http://www.usa.gov/publicdomain/label/1.0/", instrument = "Archival Tag", 
    
    Conventions = "CF-1.8", standard_name_vocabulary = "CF Standard Name Table v79",
    source = paste0("Data processed using trawllight ", packageVersion(pkg =
    "trawllight")))
)
}
\arguments{
\item{region}{Survey region as a character vector ("BS" = EBS and NBS, "GOA" = Gulf of Alaska, "AI" = Aleutian Islands)}

\item{year}{Year or years for which data product should be generated.}

\item{global_attributes}{List of global attributes that is passed to gapctd::df_to_netcdf(global_attributes).}
}
\description{
Bundle apparent optical properties data product into a netCDF file for the specified year and region.
}
