% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk9_cast_times.R
\name{mk9_cast_times}
\alias{mk9_cast_times}
\title{Get cast times}
\usage{
mk9_cast_times(channel = NULL, survey, vessel, cruise)
}
\arguments{
\item{channel}{Oracle connection as an RODBC connection object.}

\item{survey}{RACE survey code as a character vector.}

\item{vessel}{RACE vessel code a numeric vector.}

\item{cruise}{RACE cruise code as a numeric vector.}
}
\value{
Objects created: MK9.performance.csv - trawl-mounted light meter deployment performance code, CastTimes.csv - catalog of events during trawl haul operation. MK9 performance codes: 0 = good performance, 1 = bad performance, 99 = no data between on and off bottom events.
}
\description{
This cast.times function retrieves data from Oracle and offset tables to generate a table of event times defining a trawl event. These events are: downcast start, downcast end, on bottom, haul back, off bottom, upcast start and upcast end; they bracket he portion of the trawl haul event that is of interest to us for the use of MK9 irradiance data.  Within this function MK9 light meter depths and times are adjusted to using the info provided in the offsets table so that MK9 depth is calibrated to the survey MBT and MK9 clock drift is adjusted to synch up with the MBT/GPS time that is the basis for the survey clock.
}
