% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk9_convert_wcdap.R
\name{mk9_convert_wcdap_archive}
\alias{mk9_convert_wcdap_archive}
\title{Convert WCDAP archive csv to trawllight csv format}
\usage{
mk9_convert_wcdap_archive(
  dir_path = NULL,
  file_name = NULL,
  plot_check = TRUE,
  type = NULL
)
}
\arguments{
\item{dir_path}{Path to the directory containing the Archive.csv file}

\item{file_name}{Name of the Archive.csv file. If NULL, searches for the file name in dir_path.}

\item{plot_check}{Should a plot (.png) of datetime versus depth be written to dir_path?}

\item{type}{User specified type. Detected automatically if type = NULL (default). Otherwise, either "trwl" for trawl-mounted tag or "deck" for deck-mounted tag.}
}
\description{
Converts an archive CSV file produced by WC-DAP to a CSV format that is compatible with trawllight. The CSV format for trawllight is based on a format that had been produced by Wildlife Computers HexDecoder software prior to 2017
}
