% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk9_convert_wctab.R
\name{mk9_convert_wctab}
\alias{mk9_convert_wctab}
\title{Convert WC Instrument Helper file to trawllight csv format}
\usage{
mk9_convert_wctab(file_path)
}
\arguments{
\item{file_path}{Path to the file to be converted.}
}
\description{
Converts an R table file generated by WC Instrument helper to a CSV format that is compatible with trawllight.
}
