% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk9_find_surface_offset.R
\name{mk9_find_surface_offset}
\alias{mk9_find_surface_offset}
\title{Find and apply time offset to deck tag data}
\usage{
mk9_find_surface_offset(
  channel = NULL,
  survey,
  vessel,
  cruise,
  try_offsets = seq(-8, 8, 1),
  results_file = NULL
)
}
\arguments{
\item{channel}{Oracle connection as an RODBC connection object.}

\item{survey}{RACE survey code as a character vector.}

\item{vessel}{RACE vessel code a numeric vector.}

\item{cruise}{RACE cruise code as a numeric vector.}

\item{results_file}{Character vector indicating a filepath where offset results should be stored.}
}
\value{
Writes time-shifted surface values to corr_deck.csv
}
\description{
Calculates correlations between predicted and observed irradiance to calculate an archival tag offset for cases where tag time doesn't match survey time. Iterates through surface irradiance predictions from fishmethods::astrocalc4r based on location data and offset time.
}
