% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk9_setup_dir.R
\name{mk9_setup_dir}
\alias{mk9_setup_dir}
\title{Copy MK9 data and setup directory}
\usage{
mk9_setup_dir(source_path, survey, cruise, vessel, overwrite = FALSE)
}
\arguments{
\item{source_path}{Filepath to the data source}

\item{survey}{RACE survey code as a character vector.}

\item{cruise}{RACE cruise code as a numeric vector.}

\item{vessel}{RACE vessel code a numeric vector.}

\item{overwrite}{Should the local data directory be overwritten?}
}
\description{
Copy MK9 data and setup directory
}
