% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk9_time_offset.R
\name{mk9_time_offset}
\alias{mk9_time_offset}
\title{Estimate offset between MK9 and MBT times}
\usage{
mk9_time_offset(vessel, cruise, survey, channel = NULL)
}
\arguments{
\item{vessel}{RACE vessel code a numeric vector.}

\item{cruise}{RACE cruise code as a numeric vector.}

\item{survey}{RACE survey code as a character vector.}

\item{channel}{Oracle connection as an RODBC connection object.}
}
\description{
This function determines the offset in seconds between the date_times of the MK9 and those of the MBT by aligning depths during the downcasts.
}
