% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_all_surface}
\alias{process_all_surface}
\title{Loop surface_light}
\usage{
process_all_surface(dir.structure, adjust.time = T, ...)
}
\arguments{
\item{dir.structure}{A character vector containing filepaths for all of the directories containing light measurements from surface/deck archival tags.}

\item{...}{Additional arguments to be passed to surface_light or time_adjustments}
}
\description{
For use processing AOPs from AFSC/RACE/GAP data structure. Iterates surface_light over directories containing deck**.csv files and CastTimes.csv files.
}
