% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quanta_to_energy}
\alias{quanta_to_energy}
\title{Photon flux density to radiometric energy}
\usage{
quanta_to_energy(x, wavelength)
}
\arguments{
\item{x}{Numeric vector. Energy for a wavelength in units of watts.}

\item{wavelength}{Numeric vector.Wavelength in nanometers.}
}
\value{
Numeric vector of energy in radiometric energy in watts.
}
\description{
Convert quantum units of micromoles of photons per meter squared per second to radiometric energy (watts per meter squared per second)
}
\details{
Convert energy to quanta based on wavelength.
}
