% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radtran.R
\name{radtran}
\alias{radtran}
\title{RADTRAN Solar Irradiance Model with delta-Eddington cloud transmission}
\usage{
radtran(
  latitude,
  longitude,
  zenith_degrees = NULL,
  doy = NULL,
  hour = NULL,
  minute = NULL,
  alpha,
  surface_pressure,
  water_vapor = NA,
  aerosol_scale_height = NA,
  visual_range = NA,
  air_mass = NA,
  relative_humidity = NA,
  wind_speed = 0,
  cloud_modification = TRUE,
  cloud_droplet_radius = 11.8,
  liquid_water_path = 125,
  water_refraction = 1.341,
  foam_spectrum = FALSE,
  ozone = "vanheuklon",
  w_v = 350:700,
  e_v = NA
)
}
\arguments{
\item{latitude}{Latitude in decimal degrees (East is positive, West is negative)}

\item{longitude}{Longitude in decimal degrees (North is positive, South is negative)}

\item{zenith_degrees}{Solar zenith angle in degrees. Pass either externally provided zenith_degrees or hour, and minute. Default = NULL.}

\item{doy}{Day of year (numeric vector of length 1 with a values of 1-365).}

\item{hour}{Local hour. Not used of zenith_degrees is provided.}

\item{minute}{Local minute. Not used of zenith_degrees is provided.}

\item{alpha}{Power on angstrom turbidity expression}

\item{surface_pressure}{Surface pressure (milibars)}

\item{water_vapor}{precipitable water vapor (cm)}

\item{aerosol_scale_height}{Aerosol scale height in kilometers. Only used if cloud_modification = FALSE.}

\item{visual_range}{Visibility in kilometers. Only used if cloud_modification = FALSE.}

\item{air_mass}{Air mass type (1-10). Typically 1-10, where 1 correspondswith marine aerosols, 10 with continental aerosols.}

\item{relative_humidity}{Relative humidity (percent).}

\item{wind_speed}{Wind speed in meters per second. Must be greater than or equal to zero. Default = 0 (calm)}

\item{cloud_modification}{Logical. Use the cloud cover modification from Gregg (2002), based on Slingo (1989)? Default = TRUE.}

\item{cloud_droplet_radius}{Cloud droplet radius in micrometers. Used for cloud_modification. Default 11.8 micrometers = mean oceanic value from Han et al. (1994).}

\item{liquid_water_path}{Liquid water path used for cloud_modification.}

\item{water_refraction}{Refractive index of water. Default for seawater = 1.341.}

\item{foam_spectrum}{Logical. TRUE = use Gregg (2002) wavelength-dependent correction for sea-foam reflectance-based on Frouin et al. (1996). FALSE = original implementation with no spectral dependence (Gregg and Carder 1990)}

\item{ozone}{Either a 1L numeric vector indicating the total ozone value or a character vector indicating an algorithmic approximation method. The only approximation method currently available is "vanheuklon" based on Van Heuklon (1979).}

\item{w_v}{Numeric vector indicating wavelengths to use. Default set to maximum range of 350:700}

\item{e_v}{Optional. Numeric vector of top-of-atmosphere irradiances. If not provided, uses default 1-nm resolution solar TOA irradiance from Gregg and Carder (1990).}

\item{below_surface}{Logical. TRUE = estimate irradiance just below the sea surface. FALSE = estimate irradiance just above the sea surface.}
}
\value{
A data frame containing spectral direct and diffuse irradiance by wavelength at the sea surface and just below the sea surface.
}
\description{
An R-based implementation of the Gregg and Carder (1990) model for cloudless maritime atmospheres, along with an option to use the delta-Eddington approximation of two-stream approach for spectral cloud transmission following Slingo (1989). Produces estimates of direct and diffuse spectral irradiance at the sea surface and just below the sea surface at 1 nm resolution for 350-700 nm wavelengths. Atmospheric path length calculation (air mass) follows Kasten and Young (1989) for zenith angles up to 90 degrees. At zenith angles larger than 90 degrees atmospheric path lengths are calculated using Schoenberg's (1929) spherical approximation assuming an earth radius of 6371 km and atmosphere height of 9 km. Ozone can be user-specified or approximated following Van Heuklon (1979).
}
\references{
Gregg, W.W., 2002. A coupled ocean-atmosphere radiative model for global ocean biogeochemical models. NASA Global Modeling and Assimilation Series, M. Suarez (ed.), NASA Technical Memorandum 2002-104606, Volume 22.

Gregg, W.W., Carder, K.L., 1990. A simple spectral solar irradiance model for cloudless maritime atmospheres. Limnol. Oceanogr. 35, 1657–1675. https://doi.org/10.4319/lo.1990.35.8.1657

Kasten, F., 1965. A new table and approximation formula for the relative optial air mass. Arch. für Meteorol. Geophys. und Bioklimatologie Ser. B 14, 206–223. https://doi.org/10.1007/BF02248840

Kasten, F., Young, A.T., 1989. Revised optical air mass tables and approximation formula. Appl. Opt. 28, 4735–4738. https://doi.org/10.1364/AO.28.004735

Schoenberg, E. (1929). Theoretische Photometrie, g) Über die Extinktion des Lichtes in der Erdatmosphäre. In Handbuch der Astrophysik. Band II, erste Hälfte. Berlin: Springer.

Slingo, A., 1989. A GCM parameterization for the shortwave radiative properties of water clouds. J. Atmos. Sci. 46, 1419–1427. https://doi.org/10.1175/1520-0469(1989)046<1419:AGPFTS>2.0.CO;2

Van Heuklon, T.K., 1979. Estimating atmospheric ozone for solar radiation models. Sol. Energy 22, 63–68. https://doi.org/10.1016/0038-092X(79)90060-4
}
\author{
Sean Rohan \email{sean.rohan@noaa.gov}
}
