% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radtran_solar_zenith.R
\name{radtran_solar_zenith}
\alias{radtran_solar_zenith}
\title{Calculate solar zenith angle}
\usage{
radtran_solar_zenith(lat, lon, day, month, year, hour, timezone)
}
\arguments{
\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees.}

\item{day}{Day of month. Integer.}

\item{month}{Month. Integer.}

\item{year}{Year. Numeric.}

\item{hour}{Hour (decimal hour). Numeric.}

\item{timezone}{Local time relative to GMT (Alaska Daylight Time is -8, i.e., GMT-8)}
}
\value{
Returns a numeric vector of solar zenith angle (degrees).
}
\description{
Calculate solar zenith angle for a latitude and longitude based on position, date, time, and timezone. From the function fishmethods::astrocalc4r.
}
