% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{raw_light}
\alias{raw_light}
\title{Raw light}
\usage{
raw_light(
  dir.structure,
  time.buffer = 20,
  silent = T,
  column = T,
  adjust.time = T,
  ...
)
}
\arguments{
\item{dir.structure}{Filepath to directory with corr_MK9hauls.csv and CastTimes.csv files}

\item{time.buffer}{Numeric vector 1L. Buffer around haul start time.}

\item{silent}{Logical. Should messages be printed to the console during processing}

\item{adjust.time}{Should times be shifted to account for archival tags being set to the wrong time zones?}

\item{...}{Additional arguments passed to internal functions.}

\item{coulumn}{Logical. Should light measurements from casts corr_MK9hauls.csv be retrieved?}
}
\description{
For use processing AOPs from AFSC/RACE/GAP data structure.
}
