% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spctral2.R
\name{spctral2}
\alias{spctral2}
\title{Bird and Riordan Solar Irradiance Model}
\usage{
spctral2(
  latitude,
  longitude,
  doy,
  hour,
  minute,
  tz,
  alpha,
  angle_of_incidence,
  n_wavelengths,
  ozone,
  albedo,
  surface_tilt,
  surface_azimuth = 180,
  surface_pressure,
  aod,
  water_vapor,
  w_v,
  r_v,
  z,
  asym,
  omega,
  omega_p
)
}
\arguments{
\item{latitude}{Latitude in decimal degrees (East is positive, West is negative)}

\item{longitude}{Longitude in decimal degrees (North is positive, South is negative)}

\item{doy}{Day of year (numeric vector of length 1 with a values of 1-365)}

\item{hour}{Hour}

\item{minute}{Minute}

\item{tz}{Timezone}

\item{alpha}{Power on angstrom turbidity expression (default = 1.14 for rural areas}

\item{angle_of_incidence}{Angle of incidence of direct beam on flat surface (in degrees)}

\item{n_wavelengths}{Number of wavelengths}

\item{ozone}{Ozone (O_3) in atmospheres per centimeter}

\item{albedo}{Ground albedo. Default is a generic value of 0.06 for sea surface}

\item{surface_tilt}{Tilt angle of ground surface from horizontal (degrees)}

\item{surface_azimuth}{Aziumuth of surface}

\item{surface_pressure}{Surface pressure (milibars)}

\item{aod}{aerosol optical depth at 0.5 microns (base e)}

\item{water_vapor}{precipitable water vapor (cm)}

\item{w_v}{Vector of wavelengths (length the same as length of r_v)}

\item{r_v}{Vector of reflectivity at wavelengths (length the same as length of w_v)}

\item{asym}{Aerosol scattering asymmetry factor (forward to total scattering ratio)}

\item{omega}{Single scattering albedo factor}

\item{omega_p}{Single scattering albedo factor}

\item{aoi}{}

\item{zenith_degrees}{Solar zenith angle (degrees). If not provided, must provide latitude and longitude.}
}
\value{
A data frame containing spectral direct and diffuse irradiance by wavelength in units of photon flux density (photons per square meter per second)
}
\description{
An entirely R-based implementation of the Bird and Riordan (1984, 1986) SPCTRAL2 model to estimate direct and diffuse spectral irradiance on horizontal and titled surfaces at the Earth's surface. Model based on the FORTRAN program SPCTRAL2 (Bird and Riordan, 1984) with corrections described by Bird and Riordan (1986). As of July 2021, FORTRAN and C implementations of SPCTRAL2 are available from the National Renewable Energy Laboratory: [https://www.nrel.gov/grid/solar-resource/spectral.html](https://www.nrel.gov/grid/solar-resource/spectral.html)
}
\references{
Bird, R., Riordan, C. 1984. Simple solar spectral model for direct and diffuse irradiance on horizontal and titled planes at the Earth's surface for cloudless atmospheres. SERI/TR-2145-2436. Solar Energy Research Institute. 37 pp.

Bird, R.E., Riordan, C., 1986. Simple solar spectral model for direct and diffuse irradiance on horizontal and tilted planes at the Earth’s surface for cloudless atmospheres. J. Clim. Appl. Meteorol. 25, 87–97. https://doi.org/10.1175/1520-0450(1986)025<0087:SSSMFD>2.0.CO;2
}
