% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{surface_light}
\alias{surface_light}
\title{Find surface light measurements during casts}
\usage{
surface_light(
  light.data,
  cast.data,
  time.buffer = 30,
  agg.fun = trawllight::geometric.mean,
  ...
)
}
\arguments{
\item{light.data}{Light measurements from the surface/deck archival tag.}

\item{cast.data}{Haul event times indicating the start and end times for net deployment and retrival.}

\item{time.buffer}{Time buffer before and after the start of the cast.}

\item{agg.fun}{Function applied to calculate central tendency metric for light measurements sampled during the cast time window. Default trawllight::geometric.mean}

\item{...}{Additional arguments passed to internal functions.}
}
\description{
For use processing AOPs from AFSC/RACE/GAP data structure. Uses cast start and end times to find concurrent measurements obtained by the deck-mounted archival tag, including a buffer. The buffer is added around the start and end times, so a 30 second buffer = one minute.
}
