% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_residuals_indirect.R
\name{tag_residuals_indirect}
\alias{tag_residuals_indirect}
\title{Indirect method for detecting tag obstruction}
\usage{
tag_residuals_indirect(
  x,
  formula = log10(trans_llight) ~ s(PAR, bs = "cr"),
  utc.offset = NULL,
  lat.col = "latitude",
  lon.col = "longitude",
  time.col = "start_time",
  light.col = "trans_llight",
  depth.bins = c(1, 3, 5, 7, 9),
  ...
)
}
\arguments{
\item{x}{Data frame containing light at depth, depth, latitude (decimal degrees), longitude (decimal degrees), cast start time in POSIXct format, and timezone offset relative to UTC and unique identifiers for survey vessels, years, etc.}

\item{formula}{an object of the class formula, passed to the generalized additive model. See documentation for \code{\link[mgcv]{gam}}.}

\item{utc.offset}{A numeric vector providing the timezone offset relative to UTC. See documentation for \code{\link[fishmethods]{astrocalc4r}}.}

\item{lat.col}{A character vector with the name of the column containing latitude.}

\item{lon.col}{A character vector with the name of the column containing longitude.}

\item{time.col}{A character vector with the name of the column containing start time.}

\item{light.col}{A character vector with the name of the column containing water column light measurements.}

\item{...}{Additional values for gam.}
}
\value{
Returns a data frame containing input data, solar position, photosynthetically active radiation, and generalized additive model residuals.
}
\description{
\code{tag_residuals_indirect} fit a generalized additive model between surface light predictions from a global irradiance model (Frouin et al. 1989; fishmethods::astrocalc4r) and water column light, based on the position and time of the cast. Returns model residuals, estimate surface irradiance (photosynthetically active radiation), and solar position.
}
\references{
Frouin, R., Lingner, D.W., Gautier, C., Baker, K.S., and Smith, R.C. 1989. A simple analytical formula to compute clear sky total and photosynthetically available solar irradiance at the ocean surface. J. Geophys. Res. 94(C7): 9731. doi:10.1029/JC094iC07p09731.

Gary A. Nelson (2017). fishmethods: Fishery Science Methods and Models in R. R package version 1.10-4.https://CRAN.R-project.org/package=fishmethods

Wood, S.N. (2011) Fast stable restricted maximum likelihood and marginal likelihood estimation of semiparametric generalized linear models. Journal of the Royal Statistical Society (B) 73(1):3-36
}
\author{
Sean Rohan \email{sean.rohan@noaa.gov}
}
