% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_adjustments}
\alias{time_adjustments}
\title{Correct tag time in cases where offsets were incorrect}
\usage{
time_adjustments(light.data, cast.data)
}
\arguments{
\item{light.data}{Data frame with light data}

\item{cast.data}{Data frame containing case data.}
}
\description{
For use processing AOPs from AFSC/RACE/GAP data structure. Make adjustments to correct inconsistencies between tag time and survey time.
}
