% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlutilities.R
\name{tlu_calc_resids}
\alias{tlu_calc_resids}
\title{Wrapper function to calculate residuals}
\usage{
tlu_calc_resids(input, depth.bins = c(1, 3, 5), write.rds = FALSE)
}
\arguments{
\item{input}{Input data, i.e., the data frame output by trawllight::tlu_combine_casts()}

\item{depth.bins}{Depth bins to use for calculating residuals. Default c(1,3,5) described by Rohan et al. (2020) and Rohan et al. (2021).}

\item{write.rds}{If TRUE, saves output to output/tag_residuals.rds}
}
\description{
Wrapper function over trawllight::tag_residuals_indirect() and trawllight_tag_residuals_direct(), which calculate residuals for the trawllight algorithm based on RACE's data structure.
}
