% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlutilities.R
\name{tlu_calc_summary}
\alias{tlu_calc_summary}
\title{Write AOPs}
\usage{
tlu_calc_summary(survey)
}
\description{
Calculate cast-level variables (near-bottom optical depth, Z10, Z1) from filtered and interpolated data and write data output/final_stn_vars.rds; and write profile variables (Kd(Z)) that pass QA/QC checks to output/final_profile_vars.rds.
}
\details{
@param survey Survey region "BS", "NBS", "GOA", "AI", or "SLOPE".
 @noRd
}
