% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlutilities.R
\name{tlu_cast_wrapper}
\alias{tlu_cast_wrapper}
\title{Applies a function to all casts in a data frame}
\usage{
tlu_cast_wrapper(x, id.col, FUN, min.rows = 4, silent = TRUE, ...)
}
\arguments{
\item{x}{Data frame}

\item{id.col}{Identification column}

\item{FUN}{Function to apply to all id.col variables}

\item{min.rows}{Minimum number of rows necessary for the function to run}

\item{silent}{Should id variables be printed to console.}

\item{...}{Optional arguments passed to FUN.}
}
\description{
For use processing AOPs from AFSC/RACE/GAP data structure.
}
