% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlutilities.R
\name{tlu_combine_casts}
\alias{tlu_combine_casts}
\title{Combine haul data with data from upcasts, downcasts, and surface}
\usage{
tlu_combine_casts(
  haul.dat = NULL,
  surface = NULL,
  downcasts = NULL,
  upcasts = NULL
)
}
\arguments{
\item{haul.dat}{Haul data as a data frame. Default NULL searches for haul_data file in the output directory.}

\item{surface}{Surface light data as a data frame. Default NULL searches for surface data in the output directory.}

\item{downcasts}{Downcast data as a list that includes the light_data data frame. Default NULL searches for downcast data in the output directory.}

\item{upcasts}{Upcast data as a list that includes the light_Data data frame. Default NULL searaches for upcast data in the output directory.}
}
\description{
Combine haul data with light data from upcasts, downcasts, and surface.
}
