% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlutilities.R
\name{tlu_find_atten_profiles}
\alias{tlu_find_atten_profiles}
\title{Merge attenuation profiles from casts that passed QA/QC}
\usage{
tlu_find_atten_profiles(downcasts, upcasts, keep)
}
\arguments{
\item{downcasts}{Data frame containing downcast attenuation profiles}

\item{upcasts}{Data frame containing upcast attenuation profiles}

\item{keep}{Data frame with QA/QC flags}
}
\description{
Merge attenuation profiles from upcast and downcasts with a data frame where data passed QA/QC
}
