% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlutilities.R
\name{tlu_flag_orientation}
\alias{tlu_flag_orientation}
\title{Flag orientation}
\usage{
tlu_flag_orientation(
  resids,
  direct.threshold = "auto",
  indirect.threshold = "auto",
  direct.col = "direct_residual",
  indirect.col = "indirect_residual"
)
}
\arguments{
\item{resids}{Residuals data frame}

\item{direct.threshold}{Threshold for direct method residual error assignment as a numeric vector or ("auto") to use the method described in Rohan et al. (2020, 2021)}

\item{indirect.threshold}{Threshold for indirect method residual error assignment as a numeric vector or ("auto") to use the method described in Rohan et al. (2020, 2021)}

\item{direct.col}{Name of direct residual column. Default = "direct_residual"}

\item{indirect.col}{Name of indirect residual column. Default = "indirect_residual"}
}
\description{
Assign orientation using trawllight::threshold_select()
}
