% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlu_get_bottom.R
\name{tlu_get_bottom}
\alias{tlu_get_bottom}
\title{Retrieve bottom light data}
\usage{
tlu_get_bottom(
  directory_structure,
  agg_fun = mean,
  time_buffer = 30,
  survey,
  ...
)
}
\arguments{
\item{directory_structure}{File path to csv file that lists directories to be processed.}

\item{agg_fun}{Function applied to calculate central tendency metric for light measurements sampled during the cast time window. Default = mean}

\item{time_buffer}{Time buffer in seconds to add/subtract from upcast and downcast times. Default = 20}

\item{survey}{RACE survey region as a character vector ("BS", "NBS", "AI", "GOA" or "SLOPE")}
}
\description{
Get bottom light data
}
