% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlu_get_casts.R
\name{tlu_get_casts}
\alias{tlu_get_casts}
\title{Wrapper function to retrieve cast data and aggregate}
\usage{
tlu_get_casts(
  directory_structure,
  survey,
  cast.dir = "downcast",
  time.buffer = 20,
  bin.size = 2,
  bin.gap = 6,
  agg.fun = geometric.mean,
  silent = TRUE,
  ...
)
}
\arguments{
\item{directory_structure}{File path to csv file that lists directories to be processed.}

\item{survey}{RACE survey region as a character vector ("BS", "NBS", "AI", "GOA" or "SLOPE".}

\item{cast.dir}{Cast direction, either "upcast" or "downcast"}

\item{time.buffer}{Time buffer in seconds to add/subtract from upcast and downcast times. Default = 20}

\item{bin.size}{Passed to trawllight::filter_stepwise. Depth bin size for aggregating light measurements.  Default = 2}

\item{bin.gap}{Passed to trawllight::filter_stepwise. Maximum allowable gap in observations before a profile is considered to not meet continuity standards. Default = 6 (i.e., three bins if bin.size = 2)}

\item{agg.fun}{Function used to calculate summary statistic for a depth bin. Default = geometric.mean}

\item{silent}{Passed to...}

\item{...}{Optional arguments passed to filter_stepwise or calculate_attenuation.}
}
\description{
Description goes here... writes binned cast data to an rds file after interpolating missing surface values.
}
