% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlu_get_surface.R
\name{tlu_get_surface}
\alias{tlu_get_surface}
\title{Wrapper function to retrieve cast data and aggregate}
\usage{
tlu_get_surface(directory_structure, survey, ...)
}
\arguments{
\item{directory_structure}{File path to csv file that lists directories to be processed.}

\item{survey}{Survey as a character vector ("BS", "NBS", "GOA", "AI", or "SLOPE")}
}
\description{
Description goes here... writes binned cast data to an rds file after interpolating missing surface values.
}
