% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlutilities.R
\name{tlu_prep_dir_list}
\alias{tlu_prep_dir_list}
\title{Write light data directories to CSV}
\usage{
tlu_prep_dir_list(survey, light_data_root = "G:/RACE_LIGHT/LightData/Data")
}
\arguments{
\item{survey}{RACE survey code as a character vector.}

\item{light_data_root}{Root directory for light data as a character vector.}
}
\description{
Find all of the directories with light data for a region and save the list of directories to a csv file.
}
