% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlutilities.R
\name{tlu_prep_haul_data}
\alias{tlu_prep_haul_data}
\title{Get haul data and write to RDS}
\usage{
tlu_prep_haul_data(channel = NULL, survey)
}
\arguments{
\item{channel}{Oracle connection as an RODBC connection object.}

\item{survey}{RACE survey code as a character vector.}
}
\description{
Run query to retrieve haul data from RACEBASE
}
