% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlu_process_all.R
\name{tlu_process_all}
\alias{tlu_process_all}
\title{Wrapper function for filter_stepwise and calculate_attenuation using RACE data structure.}
\usage{
tlu_process_all(
  dir.structure,
  time.buffer = 20,
  cast.dir = "downcast",
  agg.fun = median,
  binsize = 2,
  bin.gap = 6,
  kz.binsize = 0.5,
  silent = TRUE,
  ...
)
}
\arguments{
\item{dir.structure}{Vector of file paths to directories containing corr_MK9hauls.csv and CastTimes.csv files.}

\item{time.buffer}{Buffer around upcast_start and downcast_start}

\item{cast.dir}{Cast direction, either 'Downcast' or 'Upcast'}

\item{agg.fun}{Function to use to calculate light for a depth bin (Default = median)}

\item{binsize}{Bin width for depth bins}

\item{bin.gap}{Maximum gap in depth bins for a cast to still be considered 'good'}

\item{kz.binsize}{Bin size for interpolation}
}
\description{
For use processing AOPs from AFSC/RACE/GAP data structure. This function is designed to work with the file structure of RACE light data to subset all of the Mk9 data obtained during upscasts and downcasts. process_all runs functions trawllight::convert_light, trawllight::filter_stepwise, and trawllight::calculate_attenuation on the data.
}
