% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlu_process_all_surface.R
\name{tlu_process_all_surface}
\alias{tlu_process_all_surface}
\title{Loop surface_light}
\usage{
tlu_process_all_surface(dir.structure, adjust.time = T, ...)
}
\arguments{
\item{dir.structure}{A character vector containing filepaths for all of the directories containing light measurements from surface/deck archival tags.}

\item{adjust.time}{Should trawllight::tlu_time_adjustments function be used to adjust surface times to match survey time.}

\item{...}{Additional arguments to be passed to surface_light or time_adjustments}
}
\description{
Calculate surface light for upcasts and downcasts.
}
