% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlutilities.R
\name{tlu_setup_dir}
\alias{tlu_setup_dir}
\title{Copy MK9 data and setup directory}
\usage{
tlu_setup_dir(
  channel = NULL,
  survey,
  light_data_root = "G:/RACE_LIGHT/LightData/Data"
)
}
\arguments{
\item{channel}{Oracle connection as an RODBC connection object.}

\item{survey}{RACE survey code as a character vector.}

\item{light_data_root}{Filepath to the data source}
}
\description{
Setup directory and retrieve data for processing MK9 data.
}
