% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlutilities.R
\name{tlu_use_casts}
\alias{tlu_use_casts}
\title{Combine orientation and quality for upcasts and downcasts}
\usage{
tlu_use_casts(input)
}
\arguments{
\item{input}{Input data frame}
}
\description{
Combine orientation and quality for upcasts and downcasts
}
