% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlutilities.R
\name{tlu_use_casts2}
\alias{tlu_use_casts2}
\title{QA/QC: Interpolated too bright}
\usage{
tlu_use_casts2(input)
}
\arguments{
\item{input}{Input data frame}
}
\description{
Omit casts where estimated near-surface light exceeds maximum observed
}
