% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlu_vertical_profiles.R
\name{tlu_vertical_profiles}
\alias{tlu_vertical_profiles}
\title{Subsets light measurements from upcasts/downcasts}
\usage{
tlu_vertical_profiles(light.data, cast.data)
}
\arguments{
\item{light.data}{Data frame with light data}

\item{cast.data}{Data frame containing case data.}
}
\description{
Assigns light measurements to upcast or downcast based on downcast start time and downcast end time.
}
