% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trawllight.R
\docType{package}
\name{trawllight}
\alias{trawllight}
\alias{trawllight-package}
\title{trawllight: A package for computating the notorious bar statistic.}
\description{
The trawllight package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

