% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_light.R
\name{convert_light}
\alias{convert_light}
\title{Convert archival tag light to PFD}
\usage{
convert_light(x)
}
\arguments{
\item{x}{Archival tag light measurement in relative units (0-256)}
}
\value{
Light in units of photon flux density.
}
\description{
\code{convert_light} converts light measurements from relative units recorded by a Wildlife Computers TDR-Mk9 archival tag to photon flux density, based on the equation of Kotwicki et al. (2009).
}
\references{
Kotwicki, S., Robertis, A., von Szalay, P., and Towler, R. 2009. The effect of light intensity on the availability of walleye pollock (Theragra chalcogramma) to bottom trawl and acoustic surveys. Can. J. Fish. Aquat. Sci. 66(6): 983–994. doi:10.1139/f09-055.
}
