% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_surface.R
\name{estimate_surface}
\alias{estimate_surface}
\title{Estimate missing light measurements}
\usage{
estimate_surface(x, atten.col = "k_linear", ratio.col = "light_ratio",
  light.col = "trans_llight", depth.col = "cdepth", id.col = c("vessel",
  "cruise", "haul", "updown", "quality", "k_column"))
}
\arguments{
\item{x}{Data frame containing depth, proportional light, binned light measurement, depth, and additional columns.}

\item{atten.col}{Name of the column containing the linear attenuation coefficient. Default = "k_linear"}

\item{ratio.col}{Name of the column containing proportional light level. Default = "light_ratio"}

\item{light.col}{Name of the column containing light measurement. Default = "trans_llight"}

\item{depth.col}{Name of the column containing depth. Default = "cdepth"}

\item{id.col}{Vector of column names which uniquely identify data for a single cast.}
}
\description{
Uses the linear attenuation coefficient of downwelling irradiance for the shallowest depth bins with 'good' quality light measurements to estimate light for the shallowest depth bin.
}
\author{
S.K. Rohan \email{skrohan@uw.edu}
}
