% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk9_extinction.R
\name{mk9_extinction}
\alias{mk9_extinction}
\title{Write time-corrected data to csv}
\usage{
mk9_extinction(channel, survey, vessel, cruise)
}
\arguments{
\item{channel}{Oracle connection as an RODBC connection object.}

\item{survey}{RACE survey code as a character vector.}

\item{vessel}{RACE vessel code a numeric vector.}

\item{cruise}{RACE cruise code as a numeric vector.}
}
\description{
Use estimated offsets to time-correct MK9 data.
}
