% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_spctral2.R
\name{spctral2}
\alias{spctral2}
\title{SPCTRAL2 Solar Irradiance Model}
\usage{
spctral2(
  latitude,
  longitude,
  doy,
  hour,
  minute,
  alpha,
  angle_of_incidence,
  ozone = NA,
  albedo,
  surface_tilt,
  surface_azimuth = 180,
  surface_pressure,
  aod,
  water_vapor,
  asym = 0.65,
  omega = 0.945,
  omega_p = 0.095,
  cloud_modification = FALSE
)
}
\arguments{
\item{latitude}{Latitude in decimal degrees (East is positive, West is negative)}

\item{longitude}{Longitude in decimal degrees (North is positive, South is negative)}

\item{doy}{Day of year (numeric vector of length 1 with a values of 1-365)}

\item{hour}{Local hour}

\item{minute}{Local minute}

\item{alpha}{Power on angstrom turbidity expression (default = 1.14 for rural areas}

\item{angle_of_incidence}{Angle of incidence of direct beam on flat surface (in degrees)}

\item{ozone}{Ozone (O_3) in atmospheres per centimeter}

\item{albedo}{Ground albedo. Default is a generic value of 0.06 for sea surface}

\item{surface_tilt}{Tilt angle of ground surface from horizontal (degrees)}

\item{surface_azimuth}{Aziumuth of surface}

\item{surface_pressure}{Surface pressure (milibars)}

\item{aod}{aerosol optical depth at 0.5 microns (base e)}

\item{water_vapor}{precipitable water vapor (cm)}

\item{asym}{Aerosol scattering asymmetry factor (forward to total scattering ratio)}

\item{omega}{Single scattering albedo at 400 nm. Default = 0.945}

\item{omega_p}{Single scattering wavelength variation factor. Default = 0.095}

\item{cloud_modification}{Logical. Should the cloud cover modification from Bird et al. (1987) be used. Default = TRUE.}
}
\value{
A data frame containing spectral direct and diffuse irradiance by wavelength at the earth's surface.
}
\description{
An R-based implementation of the Bird and Riordan (1984, 1986) SPCTRAL2 model. Produces estimates direct and diffuse spectral irradiance on horizontal and titled surfaces at the Earth's surface for 300-4000 nm wavelengths. This R implementation is based on code for the FORTRAN program SPCTRLA2, including corrections to code and equations described by Bird and Riordan (1986). As of July 2021, FORTRAN and C implementations of SPCTRAL2 using the original Bird and Riordan (1984, 1986) formulation is available from the National Renewable Energy Laboratory: [https://www.nrel.gov/grid/solar-resource/spectral.html](https://www.nrel.gov/grid/solar-resource/spectral.html).
}
\references{
Bird, R., Riordan, C. 1984. Simple solar spectral model for direct and diffuse irradiance on horizontal and titled planes at the Earth's surface for cloudless atmospheres. SERI/TR-2145-2436. Solar Energy Research Institute. 37 pp.

Bird, R.E., Riordan, C., 1986. Simple solar spectral model for direct and diffuse irradiance on horizontal and tilted planes at the Earth’s surface for cloudless atmospheres. J. Clim. Appl. Meteorol. 25, 87–97. https://doi.org/10.1175/1520-0450(1986)025<0087:SSSMFD>2.0.CO;2
}
